fix: bot configuration is updated

+++ src/helpers/functions.ts
@@ -1,8 +1,10 @@
 import prisma from '../lib/prisma';
-import axios from 'axios';
 import bot from '../bot/bot';
 import { keyboards } from '../constants/constants';
 import { KeyboardValue } from '../types/keyboards';
+import axiosbase from 'axios';
+import { v4 as uuidv4 } from 'uuid';
+import axios from '../utils/axiosinstance';
 
 const token = process.env.BOT_TOKEN;
 
@@ -13,11 +15,16 @@ export async function getFileUrlFromTelegram(fileId: string) {
   return fileUrl;
 }
 
-async function downloadFileAsBase64(fileUrl: string) {
-  const response = await axios.get(fileUrl, { responseType: 'arraybuffer' });
+async function downloadFileAsUrl(fileUrl: string) {
+  const uuid = uuidv4();
+  const response = await axiosbase.get(fileUrl, {
+    responseType: 'arraybuffer',
+  });
+
   const base64String = Buffer.from(response.data, 'binary').toString('base64');
   const dataUri = `data:image/jpeg;base64,${base64String}`;
-  return dataUri;
+  const res = await uploadFile(dataUri, uuid);
+  return res;
 }
 
 export const handleBannerInsertImage = async ({
@@ -28,14 +35,15 @@ export const handleBannerInsertImage = async ({
   banner: KeyboardValue;
 }) => {
   const fileUrl = await getFileUrlFromTelegram(fileId);
-  const base64String = await downloadFileAsBase64(fileUrl);
-  console.log('banner', banner, fileId);
+  const data = await downloadFileAsUrl(fileUrl);
+  if (!data?.file?.url) return console.log('Error: No file url');
   if (banner === keyboards.banner) {
     return await prisma.banner.deleteMany().then(async () => {
       await prisma.banner.create({
         data: {
           fileId,
-          fileBase64: base64String,
+          fileUrl: data?.file?.url,
+          fileName: data?.file?.name,
         },
       });
     });
@@ -45,7 +53,8 @@ export const handleBannerInsertImage = async ({
       await prisma.travelBanner.create({
         data: {
           fileId,
-          fileBase64: base64String,
+          fileUrl: data?.file?.url,
+          fileName: data?.file?.name,
         },
       });
     });
@@ -54,8 +63,35 @@ export const handleBannerInsertImage = async ({
     return await prisma.gallery.create({
       data: {
         fileId,
-        fileBase64: base64String,
+        fileUrl: data?.file?.url,
+        fileName: data?.file?.name,
       },
     });
   }
 };
+
+const uploadFile = async (uri: string, uuid: string) => {
+  const blob = await fetch(uri).then((res) => res.blob());
+  const file = new File([blob], uuid + '.png', { type: 'image/png' });
+  const formData = new FormData();
+  formData.append('file', file);
+  const res = await axios.post('/file/upload', formData, {
+    headers: {
+      'Content-Type': 'multipart/form-data',
+    },
+  });
+
+  return res.data;
+};
+
+// .then(async (blob) => {
+//   const file = new File([blob], 'image.png', { type: 'image/png' });
+//   const formData = new FormData();
+//   formData.append('file', file);
+//   const res = await axios.post('/file/upload', formData, {
+//     headers: {
+//       'Content-Type': 'multipart/form-data',
+//     },
+//   });
+//   console.log('res', res.data);
+// });
